==> R/1.0. Data wrangling.R <==
# ------------------------------------------------------------------------------
# Data generated by Félix P Leiva (e-mail: felixpleiva@gmail.com)
# Script created by Félix P Leiva (e-mail: felixpleiva@gmail.com)
# ------------------------------------------------------------------------------
# When using the data or code from this manuscript, please cite it as:

#Leiva FP, Boerrigter JGJ & Verberk WCEP. (2023). The role of cell size in
#shaping responses to oxygen and temperature in fruit flies. Zenodo.
#https://doi.org/10.5281/zenodo.5949048.
# ------------------------------------------------------------------------------
# Cleaning working space
rm(list=ls())
today<-format(Sys.Date(),"%Y%m%d")
# ------------------------------------------------------------------------------
# set the working directory to the folder containing this script:
setwd(dirname(rstudioapi::getActiveDocumentContext()$path))

# check directory
getwd()
# ------------------------------------------------------------------------------
#Libraries
library(dplyr)
library(xlsx)
# ------------------------------------------------------------------------------
# create a folder for the outputs produced by running this script (if it doesn't
# already exist)
if (!file.exists("../Outputs")) dir.create("../Outputs")
# '../' goes up one level from the current working directory, so this creates
# the 'Outputs' folder just outside the 'R' folder
# ------------------------------------------------------------------------------
# load data
dat<-read.xlsx("../Data/body size related traits 20191122.xlsx",sheetName = "data")
head(dat)
# ------------------------------------------------------------------------------
# See data structure and transform if necessary
str(dat)
dat$generation<-as.factor(dat$generation)#before corona measures of RU
dat$plate<-as.factor(dat$plate)
dat$well<-as.factor(dat$well)
dat$stock<-as.factor(dat$stock)
dat$sex<-as.factor(dat$sex)
dat$left_or_rigth<-as.factor(dat$left_or_rigth)
dat$temperature<-as.numeric(dat$temperature)
dat$oxygen<-as.numeric(dat$oxygen)
# ------------------------------------------------------------------------------
# create a new column to identify individual flies
dat$ID_fly<-as.factor(paste(dat$plate,dat$well,sep = "-"))
# ------------------------------------------------------------------------------
#Exclude NAs and calculate cell area and cell number for each data frame
dat=filter(dat,trichomes_number>0)
names(dat)
dat$cell_area<-dat$counting_area/dat$trichomes_number
dat$cell_number<-dat$wing_area/dat$cell_area
dat$cell_area_log<-log10(dat$cell_area)
dat$cell_number_log<-log10(dat$cell_number)
dat$wing_area_log<-log10(dat$wing_area)
dat$fresh_mass_log<-log10(dat$fresh_mass)
# ------------------------------------------------------------------------------
# Create a new column to identify the categories of cell size (see my PhD thesis
# Chapter Figure 4.2 and Chapter 7 Figure 7.1)
dat$cell_area_cat<-"large"
dat$cell_area_cat[which(dat$stock%in%c("25180","25182"))]<-"small"
dat$cell_area_cat[which(dat$stock%in%c("28141","28247"))]<-"medium"
# ------------------------------------------------------------------------------
#select columns of interest
names(dat)
dat<-dat[,c(2:4,7,8,14,18:25)]

# filter by sex
female=filter(dat,sex!="male")
male=filter(dat,sex=="male")
# ------------------------------------------------------------------------------
# count number of observations per treatment
xtabs(formula = ~ temperature + oxygen + stock, data = female)
xtabs(formula = ~ temperature + oxygen + stock, data = male)
# ------------------------------------------------------------------------------
#add  missing treatment (25°C and 21 kPa) for females of the stock "28247" from:

# Leiva FP, Santos M, Rezende E, & Verberk WCEP. (2021). Draft version of paper
# data and code of manuscript: Intraspecific variation on heat tolerance in a
# model ectotherm: effects of body mass, cell size, oxygen and sex (2.0).
# Zenodo. https://doi.org/10.5281/zenodo.5244364

# subset of females by stock
f25180<-subset(dat,dat$stock=="25180" & dat$sex=="female")
f25182<-subset(dat,dat$stock=="25182" & dat$sex=="female")
f25203<-subset(dat,dat$stock=="25203" & dat$sex=="female")
f28141<-subset(dat,dat$stock=="28141" & dat$sex=="female")
f28196<-subset(dat,dat$stock=="28196" & dat$sex=="female")
f28247<-subset(dat,dat$stock=="28247" & dat$sex=="female")

m25180<-subset(dat,dat$stock=="25180" & dat$sex=="male")
m25182<-subset(dat,dat$stock=="25182" & dat$sex=="male")
m25203<-subset(dat,dat$stock=="25203" & dat$sex=="male")
m28141<-subset(dat,dat$stock=="28141" & dat$sex=="male")
m28196<-subset(dat,dat$stock=="28196" & dat$sex=="male")
m28247<-subset(dat,dat$stock=="28247" & dat$sex=="male")

new<-data.frame("28247","25","21","female","new_fly",
                "1.465","1051254.1","162.5482","6457.345",
                "0.1658376","6.021708","3.810054",
                "2.210982","medium")

names(new)<-c("stock","temperature","oxygen","sex","ID_fly",
                "fresh_mass","wing_area","cell_area","cell_number",
                "fresh_mass_log","wing_area_log","cell_number_log",
                "cell_area_log","cell_area_cat")

#defining the data structure of the new data frame
new$stock<-as.factor(new$stock)
new$sex<-as.factor(new$sex)
new$ID_fly<-as.factor(new$ID_fly)
new$temperature<-as.numeric(new$temperature)
new$oxygen<-as.numeric(new$oxygen)
new$fresh_mass<-as.numeric(new$fresh_mass)
new$wing_area<-as.numeric(new$wing_area)
new$cell_area<-as.numeric(new$cell_area)
new$cell_number<-as.numeric(new$cell_number)
new$fresh_mass_log<-as.numeric(new$fresh_mass_log)
new$wing_area_log<-as.numeric(new$wing_area_log)
new$cell_area_log<-as.numeric(new$cell_area_log)
new$cell_number_log<-as.numeric(new$cell_number_log)

#merge data frames
f28247<-rbind(f28247,new)

# ------------------------------------------------------------------------------
# calculate the median of cell area for females and males reared under "control"
# conditions, i.e. 25°C and 21 kPa

# females
f25180.med<-aggregate(cell_area~stock, data = subset(f25180,f25180$temperature==25 & oxygen==21),median)
f25182.med<-aggregate(cell_area~stock, data = subset(f25182,f25182$temperature==25 & oxygen==21),median)
f25203.med<-aggregate(cell_area~stock, data = subset(f25203,f25203$temperature==25 & oxygen==21),median)
f28141.med<-aggregate(cell_area~stock, data = subset(f28141,f28141$temperature==25 & oxygen==21),median)
f28196.med<-aggregate(cell_area~stock, data = subset(f28196,f28196$temperature==25 & oxygen==21),median)
f28247.med<-aggregate(cell_area~stock, data = subset(f28247,f28247$temperature==25 & oxygen==21),median)

# males
m25180.med<-aggregate(cell_area~stock, data = subset(m25180,m25180$temperature==25 & oxygen==21),median)
m25182.med<-aggregate(cell_area~stock, data = subset(m25182,m25182$temperature==25 & oxygen==21),median)
m25203.med<-aggregate(cell_area~stock, data = subset(m25203,m25203$temperature==25 & oxygen==21),median)
m28141.med<-aggregate(cell_area~stock, data = subset(m28141,m28141$temperature==25 & oxygen==21),median)
m28196.med<-aggregate(cell_area~stock, data = subset(m28196,m28196$temperature==25 & oxygen==21),median)
m28247.med<-aggregate(cell_area~stock, data = subset(m28247,m28247$temperature==25 & oxygen==21),median)
# ------------------------------------------------------------------------------
# Calculate percentage of change (cell_area_delta) in cell area relative to the control (25°C and
# 21 kPa)

# females
f25180$cell_area_delta<-((f25180$cell_area/f25180.med[1,2])-1)*100
f25182$cell_area_delta<-((f25182$cell_area/f25182.med[1,2])-1)*100
f25203$cell_area_delta<-((f25203$cell_area/f25203.med[1,2])-1)*100
f28141$cell_area_delta<-((f28141$cell_area/f28141.med[1,2])-1)*100
f28196$cell_area_delta<-((f28196$cell_area/f28196.med[1,2])-1)*100
f28247$cell_area_delta<-((f28247$cell_area/f28247.med[1,2])-1)*100

# males
m25180$cell_area_delta<-((m25180$cell_area/m25180.med[1,2])-1)*100
m25182$cell_area_delta<-((m25182$cell_area/m25182.med[1,2])-1)*100
m25203$cell_area_delta<-((m25203$cell_area/m25203.med[1,2])-1)*100
m28141$cell_area_delta<-((m28141$cell_area/m28141.med[1,2])-1)*100
m28196$cell_area_delta<-((m28196$cell_area/m28196.med[1,2])-1)*100
m28247$cell_area_delta<-((m28247$cell_area/m28247.med[1,2])-1)*100
#------------------------------------------------------------------------------
# merge the data frames
male2<-rbind(m25180,m25182,m25203,m28141,m28196,m28247)
female2<-rbind(f25180,f25182,f25203,f28141,f28196,f28247)
dat2<-rbind(male2,female2)
# ------------------------------------------------------------------------------
# calculate the median of fresh mass per stock for males and females under
# "control" conditions, i.e. 25°C and 21 kPa

# males
m25180.med<-aggregate(fresh_mass~stock, data = subset(m25180,m25180$temperature==25 & oxygen==21),median)
m25182.med<-aggregate(fresh_mass~stock, data = subset(m25182,m25182$temperature==25 & oxygen==21),median)
m25203.med<-aggregate(fresh_mass~stock, data = subset(m25203,m25203$temperature==25 & oxygen==21),median)
m28141.med<-aggregate(fresh_mass~stock, data = subset(m28141,m28141$temperature==25 & oxygen==21),median)
m28196.med<-aggregate(fresh_mass~stock, data = subset(m28196,m28196$temperature==25 & oxygen==21),median)
m28247.med<-aggregate(fresh_mass~stock, data = subset(m28247,m28247$temperature==25 & oxygen==21),median)

# females
f25180.med<-aggregate(fresh_mass~stock, data = subset(f25180,f25180$temperature==25 & oxygen==21),median)
f25182.med<-aggregate(fresh_mass~stock, data = subset(f25182,f25182$temperature==25 & oxygen==21),median)
f25203.med<-aggregate(fresh_mass~stock, data = subset(f25203,f25203$temperature==25 & oxygen==21),median)
f28141.med<-aggregate(fresh_mass~stock, data = subset(f28141,f28141$temperature==25 & oxygen==21),median)
f28196.med<-aggregate(fresh_mass~stock, data = subset(f28196,f28196$temperature==25 & oxygen==21),median)
f28247.med<-aggregate(fresh_mass~stock, data = subset(f28247,f28247$temperature==25 & oxygen==21),median)

# Calculating the change in fresh mass relative to the control (25°C and 21 kPa)
# at the initial of the experiment for each one of the lines, at each generation

# females
f25180$fresh_mass_delta<-((f25180$fresh_mass/f25180.med[1,2])-1)*100
f25182$fresh_mass_delta<-((f25182$fresh_mass/f25182.med[1,2])-1)*100
f25203$fresh_mass_delta<-((f25203$fresh_mass/f25203.med[1,2])-1)*100
f28141$fresh_mass_delta<-((f28141$fresh_mass/f28141.med[1,2])-1)*100
f28196$fresh_mass_delta<-((f28196$fresh_mass/f28196.med[1,2])-1)*100
f28247$fresh_mass_delta<-((f28247$fresh_mass/f28247.med[1,2])-1)*100

# males
m25180$fresh_mass_delta<-((m25180$fresh_mass/m25180.med[1,2])-1)*100
m25182$fresh_mass_delta<-((m25182$fresh_mass/m25182.med[1,2])-1)*100
m25203$fresh_mass_delta<-((m25203$fresh_mass/m25203.med[1,2])-1)*100
m28141$fresh_mass_delta<-((m28141$fresh_mass/m28141.med[1,2])-1)*100
m28196$fresh_mass_delta<-((m28196$fresh_mass/m28196.med[1,2])-1)*100
m28247$fresh_mass_delta<-((m28247$fresh_mass/m28247.med[1,2])-1)*100

#merge the data frames
male2<-rbind(m25180,m25182,m25203,m28141,m28196,m28247)
female2<-rbind(f25180,f25182,f25203,f28141,f28196,f28247)
dat2<-rbind(male2,female2)

# ------------------------------------------------------------------------------
# calculate the median of wing per stock for males and females under "control" conditions, i.e. 25°C and 21 kPa

# males
m25180.med<-aggregate(wing_area~stock, data = subset(m25180,m25180$temperature==25 & oxygen==21),median)
m25182.med<-aggregate(wing_area~stock, data = subset(m25182,m25182$temperature==25 & oxygen==21),median)
m25203.med<-aggregate(wing_area~stock, data = subset(m25203,m25203$temperature==25 & oxygen==21),median)
m28141.med<-aggregate(wing_area~stock, data = subset(m28141,m28141$temperature==25 & oxygen==21),median)
m28196.med<-aggregate(wing_area~stock, data = subset(m28196,m28196$temperature==25 & oxygen==21),median)
m28247.med<-aggregate(wing_area~stock, data = subset(m28247,m28247$temperature==25 & oxygen==21),median)

# females
f25180.med<-aggregate(wing_area~stock, data = subset(f25180,f25180$temperature==25 & oxygen==21),median)
f25182.med<-aggregate(wing_area~stock, data = subset(f25182,f25182$temperature==25 & oxygen==21),median)
f25203.med<-aggregate(wing_area~stock, data = subset(f25203,f25203$temperature==25 & oxygen==21),median)
f28141.med<-aggregate(wing_area~stock, data = subset(f28141,f28141$temperature==25 & oxygen==21),median)
f28196.med<-aggregate(wing_area~stock, data = subset(f28196,f28196$temperature==25 & oxygen==21),median)
f28247.med<-aggregate(wing_area~stock, data = subset(f28247,f28247$temperature==25 & oxygen==21),median)

# Calculating the change in wing area relative to the control (25°C and 21 kPa)
# at the initial of the experiment for each one of the lines, at each generation

# females
f25180$wing_area_delta<-((f25180$wing_area/f25180.med[1,2])-1)*100
f25182$wing_area_delta<-((f25182$wing_area/f25182.med[1,2])-1)*100
f25203$wing_area_delta<-((f25203$wing_area/f25203.med[1,2])-1)*100
f28141$wing_area_delta<-((f28141$wing_area/f28141.med[1,2])-1)*100
f28196$wing_area_delta<-((f28196$wing_area/f28196.med[1,2])-1)*100
f28247$wing_area_delta<-((f28247$wing_area/f28247.med[1,2])-1)*100

# males
m25180$wing_area_delta<-((m25180$wing_area/m25180.med[1,2])-1)*100
m25182$wing_area_delta<-((m25182$wing_area/m25182.med[1,2])-1)*100
m25203$wing_area_delta<-((m25203$wing_area/m25203.med[1,2])-1)*100
m28141$wing_area_delta<-((m28141$wing_area/m28141.med[1,2])-1)*100
m28196$wing_area_delta<-((m28196$wing_area/m28196.med[1,2])-1)*100
m28247$wing_area_delta<-((m28247$wing_area/m28247.med[1,2])-1)*100

#merge the data frames
male2<-rbind(m25180,m25182,m25203,m28141,m28196,m28247)
female2<-rbind(f25180,f25182,f25203,f28141,f28196,f28247)
dat2<-rbind(male2,female2)

#Export data frame
setwd("../Outputs/")
write.csv(dat2,"1.1.1. Data to fit the models.csv",row.names = FALSE)
# ------------------------------------------------------------------------------
# saving session information with all packages versions for reproducibility purposes
sink("../Outputs/1.1.2. Data_wrangling_R_session.txt")
sessionInfo()
sink()
# ------------------------------------------------------------------------------
################################################################################
############################ END OF SCRIPT #####################################
################################################################################
