==> R/2.0. Data analyses_Table_1_and_Figure_2.r <==
# ------------------------------------------------------------------------------
# Data generated by Félix P Leiva (e-mail: felixpleiva@gmail.com)
# Script created by Félix P Leiva (e-mail: felixpleiva@gmail.com)
# ------------------------------------------------------------------------------
# When using the data or code from this manuscript, please cite it as:

#Leiva FP, Boerrigter JGJ & Verberk WCEP. (2022). The role of cell size in
#shaping responses to oxygen and temperature in fruit flies. Zenodo.
#https://doi.org/10.5281/zenodo.5949049.
#------------------------------------------------------------------------------
#Cleaning working space
rm(list=ls())
today<-format(Sys.Date(),"%Y%m%d")
# ------------------------------------------------------------------------------
# set the working directory to the folder containing this script:
setwd(dirname(rstudioapi::getActiveDocumentContext()$path))

# check directory
getwd()
# ------------------------------------------------------------------------------
# #Libraries
library(lme4)
library(dplyr)
library(car)
library(visreg)
library(AICcmodavg)
library(nlme)
library(lattice)
library(tidyr)
library(broom)
# ------------------------------------------------------------------------------
# create a folder for the outputs produced by running this script (if it doesn't
# already exist)
if (!file.exists("../Outputs")) dir.create("../Outputs")
# '../' goes up one level from the current working directory, so this creates
# the 'Outputs' folder just outside the 'R' folder
# ------------------------------------------------------------------------------
# load data
dat<-read.csv("../Outputs/1.1.1. Data to fit the models.csv")
# ------------------------------------------------------------------------------
# See data structure and transform if necessary
str(dat)
dat$sex<-as.factor(dat$sex)
################################################################################
# Test the influence of the random structure
################################################################################
# Modelling the random effect structure on fresh mass
lm.fm<-gls(fresh_mass_log ~ 1,data=dat, method = "REML")
lme.fm.i<-lme(fresh_mass_log ~ 1,random=~1|stock,data=dat,method = "REML")
lme.fm.s<-lme(fresh_mass_log ~ 1,random=~1 + stock |stock,data=dat,method = "REML")

anova(lm.fm,lme.fm.i)# Having "stock" as random effect in the intercept is more informative.
anova(lme.fm.i,lme.fm.s)# Having "stock" as random effect in the intercept and slope do not helps

# Modelling the random effect structure on wing area
lm.wa<-gls(wing_area_log ~ 1,data=dat, method = "REML")
lme.wa.i<-lme(wing_area_log ~ 1,random=~1|stock,data=dat,method = "REML")
lme.wa.s<-lme(wing_area_log ~ 1,random=~1 + stock |stock,data=dat,method = "REML")

anova(lm.wa,lme.wa.i)# Having "stock" as random effect in the intercept is more informative.
anova(lme.wa.i,lme.wa.s)# Having "stock" as random effect in the intercept and slope do not helps

# Modelling the random effect structure on cell area
lm.ca<-gls(cell_area_log ~ 1,data=dat, method = "REML")
lme.ca.i<-lme(cell_area_log ~ 1,random=~1|stock,data=dat,method = "REML")
lme.ca.s<-lme(cell_area_log ~ 1,random=~1 + stock |stock,data=dat,method = "REML")

anova(lm.ca,lme.ca.i)# Having "stock" as random effect in the intercept is more informative.
anova(lme.ca.i,lme.ca.s)# Having "stock" as random effect in the intercept and slope is less informative

################################################################################
# Does cell size category affect temperature-and-oxygen dependence of body size
# related traits?
################################################################################
# body mass
m1<-lmer(fresh_mass_log ~ temperature + sex + (1|stock),REML = FALSE, data=dat)
m2<-lmer(fresh_mass_log ~ oxygen + sex + (1|stock),REML = FALSE, data=dat)
m3<-lmer(fresh_mass_log ~ temperature + oxygen +  sex + (1|stock),REML = FALSE, data=dat)
m4<-lmer(fresh_mass_log ~ temperature * oxygen +  sex + (1|stock),REML = FALSE, data=dat)
m5<-lmer(fresh_mass_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock),REML = FALSE, data=dat)
m6<-lmer(fresh_mass_log ~ temperature * cell_area_cat +  sex + (1|stock),REML = FALSE, data=dat)
m7<-lmer(fresh_mass_log ~ oxygen * cell_area_cat +  sex + (1|stock),REML = FALSE, data=dat)


# wing area
w1<-lmer(wing_area_log ~ temperature + sex + (1|stock),REML = FALSE, data=dat)
w2<-lmer(wing_area_log ~ oxygen + sex + (1|stock),REML = FALSE, data=dat)
w3<-lmer(wing_area_log ~ temperature + oxygen +  sex + (1|stock),REML = FALSE, data=dat)
w4<-lmer(wing_area_log ~ temperature * oxygen +  sex + (1|stock),REML = FALSE, data=dat)
w5<-lmer(wing_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock),REML = FALSE, data=dat)
w6<-lmer(wing_area_log ~ temperature * cell_area_cat +  sex + (1|stock),REML = FALSE, data=dat)
w7<-lmer(wing_area_log ~ oxygen * cell_area_cat +  sex + (1|stock),REML = FALSE, data=dat)

# cell area
c1<-lmer(cell_area_log ~ temperature + sex + (1|stock),REML = FALSE, data=dat)
c2<-lmer(cell_area_log ~ oxygen + sex + (1|stock),REML = FALSE, data=dat)
c3<-lmer(cell_area_log ~ temperature + oxygen +  sex + (1|stock),REML = FALSE, data=dat)
c4<-lmer(cell_area_log ~ temperature * oxygen +  sex + (1|stock),REML = FALSE, data=dat)
c5<-lmer(cell_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock),REML = FALSE, data=dat)
c6<-lmer(cell_area_log ~ temperature * cell_area_cat +  sex + (1|stock),REML = FALSE, data=dat)
c7<-lmer(cell_area_log ~ oxygen * cell_area_cat +  sex + (1|stock),REML = FALSE, data=dat)

#-------------------------------------------------------------------------------
#Table 1
#-------------------------------------------------------------------------------
# Body mass
fit.list.table.1.m <- list(m1,m2,m3,m4,m5,m6,m7)

fit.names.table.1.m <-c("temp + sex",
                      "O2 + sex",  
                      "temp + O2 + sex",
                      "temp * O2 + sex",
                      "temp * O2 * CScat + sex",
                      "temp * CScat + sex",
                      "O2 * CScat + sex"
)

#compare by using AICc
fit.table.1.m<-aictab(fit.list.table.1.m,fit.names.table.1.m, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.table.1.m

# Wing area
fit.list.table.1.w <- list(w1,w2,w3,w4,w5,w6,w7)

fit.names.table.1.w <-c("temp + sex",
                        "O2 + sex",  
                        "temp + O2 + sex",
                        "temp * O2 + sex",
                        "temp * O2 * CScat + sex",
                        "temp * CScat + sex",
                        "O2 * CScat + sex"
)

#compare by using AICc
fit.table.1.w<-aictab(fit.list.table.1.w,fit.names.table.1.w, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.table.1.w

# Cell area
fit.list.table.1.c <- list(c1,c2,c3,c4,c5,c6,c7)

fit.names.table.1.c <-c("temp + sex",
                        "O2 + sex",  
                        "temp + O2 + sex",
                        "temp * O2 + sex",
                        "temp * O2 * CScat + sex",
                        "temp * CScat + sex",
                        "O2 * CScat + sex"
)

#compare by using AICc
fit.table.1.c<-aictab(fit.list.table.1.c,fit.names.table.1.c, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.table.1.c

# Summary Tables of the best models to be included as supplementary information

setwd("../Outputs/")#Set directory


# Export tables to .csv
write.csv(fit.table.1.m,"2.3.1. Table_1_Model_comparison_for_body_mass.csv",row.names = FALSE)
write.csv(fit.table.1.w,"2.3.2. Table_1_Model_comparison_for_wing_area.csv",row.names = FALSE)
write.csv(fit.table.1.c,"2.3.3. Table_1_Model_comparison_for_cell_area.csv",row.names = FALSE)

#-------------------------------------------------------------------------------
# Testing more complex structures on the most informative model
#-------------------------------------------------------------------------------
# fresh mass
m5<-lmer(fresh_mass_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock), data=dat,REML = FALSE)
m5.a<-lmer(fresh_mass_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock) + (1|temperature), data=dat)
m5.b<-lmer(fresh_mass_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock) + (1|oxygen),data=dat)
m5.c<-lmer(fresh_mass_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock) + (1|temperature) + (1|oxygen), data=dat,REML = FALSE)
m5.d<-lmer(fresh_mass_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock) + (1|temperature:oxygen),data=dat)
m5.e<-lmer(fresh_mass_log ~ temperature * oxygen * cell_area_cat +  sex + (cell_area_cat|stock),data=dat,REML = FALSE)
m5.f<-lmer(fresh_mass_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock) + (cell_area_cat|stock),data=dat)
m5.g<-lmer(fresh_mass_log ~ temperature * oxygen * cell_area_cat +  sex + (1+ cell_area_cat|stock),data=dat,REML = FALSE)

# we see several models DO NOT converges, but two of them does (model m5.c and m5.e and m5.g).
anova(m5.c,m5.e)
anova(m5.c,m5.g)
anova(m5,m5.c)
# more complex random structures does NOT help to explain the variation in fresh mass

# wing area
w5<-lmer(wing_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock), data=dat)
w5.a<-lmer(wing_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock) + (1|temperature), data=dat)
w5.b<-lmer(wing_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock) + (1|oxygen), data=dat)
w5.c<-lmer(wing_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock) + (1|temperature) + (1|oxygen), data=dat)
w5.d<-lmer(wing_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock) + (1|temperature:oxygen), data=dat)
w5.e<-lmer(wing_area_log ~ temperature * oxygen * cell_area_cat +  sex + (cell_area_cat|stock),data=dat)
w5.f<-lmer(wing_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock) + (cell_area_cat|stock),data=dat)
w5.g<-lmer(wing_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1+ cell_area_cat|stock),data=dat)

# we see several models DO NOT converges, but one of them does (model w5.c and w5.f).
anova(w5.c,w5.f)
anova(w5,w5.c)
# more complex random structures does not help to explain the variation in wing area

# cell area
c5<-lmer(cell_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock), data=dat)
c5.a<-lmer(cell_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock) + (1|temperature), data=dat)
c5.b<-lmer(cell_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock) + (1|oxygen), data=dat)
c5.c<-lmer(cell_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock) + (1|temperature) + (1|oxygen), data=dat)
c5.d<-lmer(cell_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock) + (1|temperature:oxygen), data=dat)
c5.e<-lmer(cell_area_log ~ temperature * oxygen * cell_area_cat +  sex + (cell_area_cat|stock),data=dat)
c5.f<-lmer(cell_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock) + (cell_area_cat|stock),data=dat)
c5.g<-lmer(cell_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1+ cell_area_cat|stock),data=dat)

# Two models DO NOT converges (model c5.e and c5.g), the rest does.
anova(c5,c5.a)
anova(c5,c5.b)
anova(c5,c5.c)
anova(c5,c5.d)
anova(c5,c5.f)

# More complex random structures does not help to explain the variation in cell area
#-------------------------------------------------------------------------------
# Summary Tables of the best models (Tables S1-S3, Supplementary information)
#-------------------------------------------------------------------------------
# fresh mass
m5<-lmer(fresh_mass_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock), data=dat)
Anova(m5.reml,test.statistic="F")
summary(m5)
coef_m<-coef(m5)
dotplot(ranef(m5,condVar=TRUE))

# wing area
w5<-lmer(wing_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock), data=dat)
Anova(w5,test.statistic="F")
summary(w5)
coef_w<-coef(w5)
dotplot(ranef(w5,condVar=TRUE))

# cell area
c5<-lmer(cell_area_log ~ temperature * oxygen * cell_area_cat +  sex + (1|stock), data=dat)
Anova(c5,test.statistic="F")
summary(c5)
coef_c<-coef(c5)
dotplot(ranef(c5,condVar=TRUE))

# Table S1-S3
Table_S1 <- capture.output(summary(m5))
Table_S2 <- capture.output(summary(w5))
Table_S3 <- capture.output(summary(c5))

write.table(Table_S1,"2.3.4. Table_S1_Estimates for the model with the highest support in Table 1_fresh mass.txt",row.names = F,quote = F)
write.table(Table_S2,"2.3.5. Table_S2_Estimates for the model with the highest support in Table 1_wing area.txt",row.names = F,quote = F)
write.table(Table_S3,"2.3.6. Table_S3_Estimates for the model with the highest support in Table 1_cell_area.txt",row.names = F,quote = F)

################################################################################
# Figure 2
setwd("../Outputs/")

# Extract general and random intercepts for fresh mass model
m_inter_general <- summary(m5)$coefficients[1, 1]
m_int_25180     <- ranef(m5)$stock[1,1]# small
m_int_25182     <- ranef(m5)$stock[2,1]# small
m_int_28141     <- ranef(m5)$stock[4,1]# medium
m_int_28247     <- ranef(m5)$stock[6,1]# medium
m_int_28196     <- ranef(m5)$stock[5,1]# large
m_int_25203     <- ranef(m5)$stock[3,1]# large

# Extract slopes per predictor for fresh mass model
m_slope_temp     <- summary(m5)$coefficients[2, 1]
m_slope_oxy      <- summary(m5)$coefficients[3, 1]
m_slope_sex      <- summary(m5)$coefficients[6, 1]
m_slope_temp_oxy <- summary(m5)$coefficients[7, 1]

#-------------------------------------------------------------------------------
# TESTS ON 20230127

# labels
labs= c("0.4","0.6","0.8","1.0","1.2")
xs= log10(c(0.4,0.6,0.8,1.0,1.2))

visreg(m5,'temperature', by='cell_area_cat',cond = list(oxygen=10),
       print.cond=TRUE,overlay=TRUE,jitter=0.1,legend = FALSE,
       points=list(col=c('#B8860B70', '#6495ED70','#B2ABD270'),cex=1.5,pch=16),
       line=list(col=c('#B8860B', '#6495ED','#B2ABD2'),lwd=4),
       xlab="Temperature (°C)", 
       #ylab=expression(paste(log[10],"[fresh mass (mg)]")),
       ylab="Fresh mass (mg)",
       yaxt="n",
       xaxt="n",
       ylim=c(-0.4,0.1),
       xlim=c(14,28),
       cex.axis=1.2,cex.lab=1.4)
#update axis
axis(1,at=c(17,25),
     labels=c("17","25"),cex.axis=1.2)
axis(2, at=xs, labels=labs, cex.axis=1.2,las=1)
# Add a legend
legend("top", legend=c("Hypoxia (10 kPa)"), cex=1.4,bty="n")
legend("bottomleft", legend=c("large", "medium","small"),
       lty = c(1,1,1),lwd=3,bty="n",cex=1.2, 
       col=c('#B8860B', '#6495ED','#B2ABD2'))
legend("topleft", legend=c("(a)"), cex=1.4,bty="n")

# add lines for each stock
x <- c(17,25)
y <- c(m_inter_general + m_slope_temp * x + m_slope_oxy * 10 + m_slope_sex + m_slope_temp_oxy * 10 * x)
lines(x,10^y,col="black")
# lines(x,10^(y + m_int_25180), col='#B2ABD250',lty=2,lwd=4) #small 25180
# lines(x,y + m_int_25182, col='#B2ABD250',lty=2,lwd=4) #small 25182
# lines(x,y + m_int_28141, col='#6495ED50',lty=2,lwd=4) #small 28141
# lines(x,y + m_int_28247, col='#6495ED50',lty=2,lwd=4) #small 28247
lines(x,10^(y + m_int_28196), col="black",lty=2,lwd=4) #small 28196
lines(x,10^(y + m_int_25203), col='#B8860B50',lty=2,lwd=4) #small 25203

lines(x,(10^y),col="black")


#-------------------------------------------------------------------------------











# Extract general and random intercepts for wing area model
w_inter_general <- summary(w5)$coefficients[1, 1]
w_int_25180     <- coef_w$stock[1,1]# small
w_int_25182     <- coef_w$stock[2,1]# small
w_int_28141     <- coef_w$stock[4,1]# medium
w_int_28247     <- coef_w$stock[6,1]# medium
w_int_28196     <- coef_w$stock[5,1]# large
w_int_25203     <- coef_w$stock[3,1]# large

# Extract slopes per predictor for wing area model
w_slope_temp     <- summary(w5)$coefficients[2, 1]
w_slope_oxy      <- summary(w5)$coefficients[3, 1]
w_slope_sex      <- summary(w5)$coefficients[6, 1]
w_slope_temp_oxy <- summary(w5)$coefficients[7, 1]

# Extract general and random intercepts for cell area model
c_inter_general <- summary(c5)$coefficients[1, 1]
c_int_25180     <- coef_c$stock[1,1]# small
c_int_25182     <- coef_c$stock[2,1]# small
c_int_28141     <- coef_c$stock[4,1]# medium
c_int_28247     <- coef_c$stock[6,1]# medium
c_int_28196     <- coef_c$stock[5,1]# large
c_int_25203     <- coef_c$stock[3,1]# large

# Extract slopes per predictor for cell area model
c_slope_temp     <- summary(c5)$coefficients[2, 1]
c_slope_oxy      <- summary(c5)$coefficients[3, 1]
c_slope_sex      <- summary(c5)$coefficients[6, 1]
c_slope_temp_oxy <- summary(c5)$coefficients[7, 1]

#{
#pdf("2.2.1. Figure 2.pdf",width = 7,height = 10,useDingbats = FALSE)
#png("2.2.1. Figure 2.png",width = 7,height = 10,units = "in",res = 600)
par(mfrow=c(3,2),tcl=-0.4, family="serif",omi=c(0,0,0,0.1))
#-------------------------------------------------------------------------------
# labels
labs= c("0.4","0.6","0.8","1.0","1.2")
xs= log10(c(0.4,0.6,0.8,1.0,1.2))

#Fresh mass
par(mai=c(0.85,0.82,0,0))
visreg(m5,'temperature', by='cell_area_cat',cond = list(oxygen=10),
           print.cond=TRUE,overlay=TRUE,jitter=0.1,legend = FALSE,
           points=list(col=c('#B8860B70', '#6495ED70','#B2ABD270'),cex=1.5,pch=16),
           line=list(col=c('#B8860B', '#6495ED','#B2ABD2'),lwd=4),
           xlab="Temperature (°C)", 
           #ylab=expression(paste(log[10],"[fresh mass (mg)]")),
           ylab="Fresh mass (mg)",
           yaxt="n",
           xaxt="n",
           ylim=c(-0.4,0.1),
           xlim=c(14,28),
           cex.axis=1.2,cex.lab=1.4)

#update axis
axis(1,at=c(17,25), 
     labels=c("17","25"),cex.axis=1.2)
axis(2, at=xs, labels=labs, cex.axis=1.2,las=1)
# Add a legend
legend("top", legend=c("Hypoxia (10 kPa)"), cex=1.4,bty="n")
legend("bottomleft", legend=c("large", "medium","small"),
       lty = c(1,1,1),lwd=3,bty="n",cex=1.2, 
       col=c('#B8860B', '#6495ED','#B2ABD2'))
legend("topleft", legend=c("(a)"), cex=1.4,bty="n")

# add lines for each stock
x <- c(17,25)
y <- c(m_inter_general + m_slope_temp * x + m_slope_oxy * 10 + m_slope_sex + m_slope_temp_oxy * 10 * x)

lines(x,y + m_int_25180, col='#B2ABD250',lty=2,lwd=4) #small 25180
lines(x,y + m_int_25182, col='#B2ABD250',lty=2,lwd=4) #small 25182
lines(x,y + m_int_28141, col='#6495ED50',lty=2,lwd=4) #small 28141
lines(x,y + m_int_28247, col='#6495ED50',lty=2,lwd=4) #small 28247
lines(x,y + m_int_28196, col='#B8860B50',lty=2,lwd=4) #small 28196
lines(x,y + m_int_25203, col='#B8860B50',lty=2,lwd=4) #small 25203
#-------------------------------------------------------------------------------
par(mai=c(0.85,0.82,0,0))
visreg(m5,'temperature', by='cell_area_cat',cond = list(oxygen=21),
       print.cond=TRUE,overlay=TRUE,jitter=0.1,legend = FALSE,
       points=list(col=c('#B8860B70', '#6495ED70','#B2ABD270'),cex=1.5,pch=16),
       line=list(col=c('#B8860B', '#6495ED','#B2ABD2'),lwd=4),
       xlab="Temperature (°C)", 
       #ylab=expression(paste(log[10],"[fresh mass (mg)]")),
       ylab="Fresh mass (mg)",
       yaxt="n",
       xaxt="n",
       ylim=c(-0.4,0.1),
       xlim=c(14,28),
       cex.axis=1.2,cex.lab=1.4)
#update axis
axis(1,at=c(17,25), 
     labels=c("17","25"),cex.axis=1.2)
axis(2, at=xs, labels=labs, cex.axis=1.2,las=1)

# Add a legend
legend("top", legend=c("Normoxia (21 kPa)"), cex=1.4,bty="n")
legend("bottomleft", legend=c("large", "medium","small"),
       lty = c(1,1,1),lwd=3,bty="n",cex=1.2, 
       col=c('#B8860B', '#6495ED','#B2ABD2'))
legend("topleft", legend=c("(b)"), cex=1.4,bty="n")

# add lines for each stock
x <- c(17,25)
y <- c(m_inter_general + m_slope_temp * x + m_slope_oxy * 21 + m_slope_sex + m_slope_temp_oxy * 21 * x)

lines(x,y + m_int_25180, col='#B2ABD250',lty=2,lwd=4) #small 25180
lines(x,y + m_int_25182, col='#B2ABD250',lty=2,lwd=4) #small 25182
lines(x,y + m_int_28141, col='#6495ED50',lty=2,lwd=4) #small 28141
lines(x,y + m_int_28247, col='#6495ED50',lty=2,lwd=4) #small 28247
lines(x,y + m_int_28196, col='#B8860B50',lty=2,lwd=4) #small 28196
lines(x,y + m_int_25203, col='#B8860B50',lty=2,lwd=4) #small 25203
#-------------------------------------------------------------------------------
#Wing area
# labels
labs= c("5","7.5","10","12.5","15")
xs= log10(c(500000,750000,1000000,1250000,1500000))

par(mai=c(0.85,0.82,0,0))
visreg(w5,'temperature', by='cell_area_cat',cond = list(oxygen=10),
       print.cond=TRUE,overlay=TRUE,jitter=0.1,legend = FALSE,
       points=list(col=c('#B8860B70', '#6495ED70','#B2ABD270'),cex=1.5,pch=16),
       line=list(col=c('#B8860B', '#6495ED','#B2ABD2'),lwd=4),
       xlab="Temperature (°C)", 
       #ylab=expression(paste(log[10], "[wing area (",mu, m^2,")]")),
       ylab=expression(paste("Wing area (",mu, m^2,"× 100,000)")),
       yaxt="n",
       xaxt="n",
       ylim=c(5.7,6.2),
       xlim=c(14,28),
       cex.axis=1.2,cex.lab=1.4)
#update axis
axis(1,at=c(17,25), 
     labels=c("17","25"),cex.axis=1.2)
axis(2, at=xs, labels=labs, cex.axis=1.2,las=1)

# Add a legend
legend("top", legend=c("Hypoxia (10 kPa)"), cex=1.4,bty="n")
legend("bottomleft", legend=c("large", "medium","small"),
       lty = c(1,1,1),lwd=3,bty="n",cex=1.2, 
       col=c('#B8860B', '#6495ED','#B2ABD2'))
legend("topleft", legend=c("(c)"), cex=1.4,bty="n")

par(mai=c(0.85,0.85,0,0))
visreg(w5,'temperature', by='cell_area_cat',cond = list(oxygen=21),
       print.cond=TRUE,overlay=TRUE,jitter=0.1,legend = FALSE,
       points=list(col=c('#B8860B70', '#6495ED70','#B2ABD270'),cex=1.5,pch=16),
       line=list(col=c('#B8860B', '#6495ED','#B2ABD2'),lwd=4),
       xlab="Temperature (°C)", 
       #ylab=expression(paste(log[10], "[wing area (",mu, m^2,")]")),
       ylab=expression(paste("Wing area (",mu, m^2,"× 100,000)")),
       yaxt="n",
       xaxt="n",
       ylim=c(5.7,6.2),
       xlim=c(14,28),
       cex.axis=1.2,cex.lab=1.4)
#update axis
axis(1,at=c(17,25), 
     labels=c("17","25"),cex.axis=1.2)
axis(2, at=xs, labels=labs, cex.axis=1.2,las=1)

# Add a legend
legend("top", legend=c("Normoxia (21 kPa)"), cex=1.4,bty="n")
legend("bottomleft", legend=c("large", "medium","small"),
       lty = c(1,1,1),lwd=3,bty="n",cex=1.2, 
       col=c('#B8860B', '#6495ED','#B2ABD2'))
legend("topleft", legend=c("(d)"), cex=1.4,bty="n")
#-------------------------------------------------------------------------------
#Cell area
labs= c("80","105","135","170","220")
xs= log10(c(80,105,135,170,220))

par(mai=c(0.85,0.82,0,0))
visreg(c5,'temperature', by='cell_area_cat',cond = list(oxygen=10),
       print.cond=TRUE,overlay=TRUE,jitter=0.1,legend = FALSE,
       points=list(col=c('#B8860B70', '#6495ED70','#B2ABD270'),cex=1.5,pch=16),
       line=list(col=c('#B8860B', '#6495ED','#B2ABD2'),lwd=4),
       xlab="Temperature (°C)", 
       #ylab=expression(paste(log[10], "[cell area (",mu, m^2,")]")),
       ylab=expression(paste("Cell area (",mu, m^2,")")),
       yaxt="n",
       xaxt="n",
       ylim=c(1.9,2.4),
       xlim=c(14,28),
       cex.axis=1.2,cex.lab=1.4)
#update axis
axis(1,at=c(17,25), 
     labels=c("17","25"),cex.axis=1.2)
axis(2, at=xs, labels=labs, cex.axis=1.2,las=1)

# Add a legend
legend("top", legend=c("Hypoxia (10 kPa)"), cex=1.4,bty="n")
legend("bottomleft", legend=c("large", "medium","small"),
       lty = c(1,1,1),lwd=3,bty="n",cex=1.2, 
       col=c('#B8860B', '#6495ED','#B2ABD2'))
legend("topleft", legend=c("(e)"), cex=1.4,bty="n")

par(mai=c(0.85,0.82,0,0))
visreg(c5,'temperature', by='cell_area_cat',cond = list(oxygen=21),
       print.cond=TRUE,overlay=TRUE,jitter=0.1,legend = FALSE,
       points=list(col=c('#B8860B70', '#6495ED70','#B2ABD270'),cex=1.5,pch=16),
       line=list(col=c('#B8860B', '#6495ED','#B2ABD2'),lwd=4),
       xlab="Temperature (°C)", 
       #ylab=expression(paste(log[10], "[cell area (",mu, m^2,")]")),
       ylab=expression(paste("Cell area (",mu, m^2,")")),
       yaxt="n",
       xaxt="n",
       ylim=c(1.9,2.4),
       xlim=c(14,28),
       cex.axis=1.2,cex.lab=1.4)
#update axis
axis(1,at=c(17,25), 
     labels=c("17","25"),cex.axis=1.2)
axis(2, at=xs, labels=labs, cex.axis=1.2,las=1)


# Add a legend
legend("top", legend=c("Normoxia (21 kPa)"), cex=1.4,bty="n")
legend("bottomleft", legend=c("large", "medium","small"),
       lty = c(1,1,1),lwd=3,bty="n",cex=1.2, 
       col=c('#B8860B', '#6495ED','#B2ABD2'))
legend("topleft", legend=c("(f)"), cex=1.4,bty="n")
#dev.off()
#}

# ------------------------------------------------------------------------------
# saving session information with all packages versions for reproducibility purposes
sink("../Outputs/2.1.1. Data analyses_Table_1_and_Figure_2_R_session.txt")
sessionInfo()
sink()
# ------------------------------------------------------------------------------
################################################################################
############################ END OF SCRIPT #####################################
################################################################################
