==> R/3.0. Data analyses_Table_2_and_Figure_3_and_4.R <==
# ------------------------------------------------------------------------------
# Data generated by Félix P Leiva (e-mail: felixpleiva@gmail.com)
# Script created by Félix P Leiva (e-mail: felixpleiva@gmail.com)
# ------------------------------------------------------------------------------
# When using the data or code from this manuscript, please cite it as:

#Leiva FP, Boerrigter JGJ & Verberk WCEP. (2023). The role of cell size in
#shaping responses to oxygen and temperature in fruit flies. Zenodo.
#https://doi.org/10.5281/zenodo.5949048.
#------------------------------------------------------------------------------
#Cleaning working space
rm(list=ls())
today<-format(Sys.Date(),"%Y%m%d")
# ------------------------------------------------------------------------------
# set the working directory to the folder containing this script:
setwd(dirname(rstudioapi::getActiveDocumentContext()$path))

# check directory
getwd()
# ------------------------------------------------------------------------------
#Libraries
library(dplyr)
library(xlsx)
library(ggplot2)
library(nlme)
library(car)
library(lmerTest)# P-values for mixed models
library(visreg)# plotting regression models
library(AICcmodavg)
library(cowplot)
library(scales)
# ------------------------------------------------------------------------------
# create a folder for the outputs produced by running this script (if it doesn't
# already exist)
if (!file.exists("../Outputs")) dir.create("../Outputs")
# '../' goes up one level from the current working directory, so this creates
# the 'Outputs' folder just outside the 'R' folder
# ------------------------------------------------------------------------------
# load data
dat<-read.csv("../Outputs/1.1.1. Data to fit the models.csv")
# ------------------------------------------------------------------------------
# See data structure and transform if necessary
str(dat)
dat$stock<-as.factor(dat$stock)
# ------------------------------------------------------------------------------
# filter by sex
female=filter(dat,sex!="male")
male=filter(dat,sex=="male")
# ------------------------------------------------------------------------------
# Modelling the random effect structure on fresh mass for females
f.lm.fm<-gls(fresh_mass~1,data=female, method = "ML")
f.lme.fm<-lme(fresh_mass~1,random=~1|stock,data=female,method = "ML")
anova(f.lm.fm,f.lme.fm)# Having "stock" as random effect in the intercept is more informative.

# Modelling the random effect structure on fresh mass for males
m.lm.fm<-gls(fresh_mass~1,data=male, method = "ML")
m.lme.fm<-lme(fresh_mass~1,random=~1|stock,data=male,method = "ML")
anova(m.lm.fm,m.lme.fm)# Having "stock" as random effect in the intercept is more informative.

# Modelling the random effect structure on fresh mass for females and males pooled
p.lm.fm<-gls(fresh_mass~1,data=dat, method = "ML")
p.lme.fm<-lme(fresh_mass~1,random=~1|stock,data=dat,method = "ML")
anova(p.lm.fm,p.lme.fm)# Having "stock" as random effect in the intercept is more informative.
#-------------------------------------------------------------------------------
# Modelling the random effect structure on wing area for females
f.lm.ws<-gls(wing_area~1,data=female, method = "ML")
f.lme.ws<-lme(wing_area~1,random=~1|stock,data=female,method = "ML")
anova(f.lm.ws,f.lme.ws)# Having "stock" as random effect in the intercept is more informative.

# Modelling the random effect structure on wing area for males
m.lm<-gls(wing_area~1,data=male, method = "ML")
m.lme<-lme(wing_area~1,random=~1|stock,data=male,method = "ML")
anova(m.lm,m.lme)# Having "stock" as random effect in the intercept is more informative.

# Modelling the random effect structure on wing area for females and males pooled
p.lm<-gls(wing_area~1,data=dat, method = "ML")
p.lme<-lme(wing_area~1,random=~1|stock,data=dat,method = "ML")
anova(p.lm,p.lme)# Having "stock" as random effect in the intercept is more informative.
#-------------------------------------------------------------------------------
# Does changes in body mass driven by cell area or cell number?
#------------------------
#FEMALES
#------------------------
#random intercept
f.lme1<-lme(fresh_mass~cell_area,random=~1|stock,data=female,method="ML")
summary(f.lme1)
#random intercepts and slopes
f.lme2<-lme(fresh_mass~cell_area,random=~1+cell_area|stock,data=female,method="ML")
summary(f.lme2)
#compare fits
AIC(f.lme1,f.lme2);anova(f.lme1,f.lme2)
#random intercepts and slopes are better

f.lme3<-lme(fresh_mass~cell_number,random=~1|stock,data=female,method="ML")
summary(f.lme3)
f.lme4<-lme(fresh_mass~cell_number,random=~1+cell_number|stock,data=female,method="ML")
summary(f.lme4)
#compare fits
AIC(f.lme3,f.lme4);anova(f.lme3,f.lme4)
#random intercepts are better

# Does changes in wing area are driven by cell area or cell number?

#random intercept
f.lme5<-lme(wing_area~cell_area,random=~1|stock,data=female,method="ML")
summary(f.lme5)
#random intercepts and slopes
f.lme6<-lme(wing_area~cell_area,random=~1+cell_area|stock,data=female,method="ML")
summary(f.lme6)
#compare fits
AIC(f.lme5,f.lme6);anova(f.lme5,f.lme6)
#random intercepts are better

f.lme7<-lme(wing_area~cell_number,random=~1|stock,data=female,method="ML")
summary(f.lme7)
f.lme8<-lme(wing_area~cell_number,random=~1+cell_number|stock,data=female,method="ML")
summary(f.lme8)
#compare fits
AIC(f.lme7,f.lme8);anova(f.lme7,f.lme8)
#random intercepts are better

#------------------------
#MALES
#------------------------
#random intercept
m.lme1<-lme(fresh_mass~cell_area,random=~1|stock,data=male,method="ML")
summary(m.lme1)
#random intercepts and slopes
m.lme2<-lme(fresh_mass~cell_area,random=~1+cell_area|stock,data=male,method="ML")
summary(m.lme2)
#compare fits
AIC(m.lme1,m.lme2);anova(m.lme1,m.lme2)
#random intercepts are better

m.lme3<-lme(fresh_mass~cell_number,random=~1|stock,data=male,method="ML")
summary(m.lme3)
m.lme4<-lme(fresh_mass~cell_number,random=~1+cell_number|stock,data=male,method="ML")
summary(m.lme4)
#compare fits
AIC(m.lme3,m.lme4);anova(m.lme3,m.lme4)
#random intercepts are better

# Does changes in wing area are driven by cell area or cell number?

#random intercept
m.lme5<-lme(wing_area~cell_area,random=~1|stock,data=male,method="ML")
summary(m.lme5)
#random intercepts and slopes
m.lme6<-lme(wing_area~cell_area,random=~1+cell_area|stock,data=male,method="ML")
summary(m.lme6)
#compare fits
AIC(m.lme5,m.lme6);anova(m.lme5,m.lme6)
#random intercepts are better

m.lme7<-lme(wing_area~cell_number,random=~1|stock,data=male,method="ML")
summary(m.lme7)
m.lme8<-lme(wing_area~cell_number,random=~1+cell_number|stock,data=male,method="ML")
summary(m.lme8)
#compare fits
AIC(m.lme7,m.lme8);anova(m.lme7,m.lme8)
#random intercepts are better

#------------------------
#SEXED POOLED
#------------------------
#random intercept
p.lme1<-lme(fresh_mass~cell_area,random=~1|stock,data=dat,method="ML")
summary(p.lme1)
#random intercepts and slopes
p.lme2<-lme(fresh_mass~cell_area,random=~1+cell_area|stock,data=dat,method="ML")
summary(p.lme2)
#compare fits
AIC(p.lme1,p.lme2);anova(p.lme1,p.lme2)
#random intercepts are better

p.lme3<-lme(fresh_mass~cell_number,random=~1|stock,data=dat,method="ML")
summary(p.lme3)
p.lme4<-lme(fresh_mass~cell_number,random=~1+cell_number|stock,data=dat,method="ML")
summary(p.lme4)
#compare fits
AIC(p.lme3,p.lme4);anova(p.lme3,p.lme4)
#random intercepts are better

# Does changes in wing area are driven by cell area or cell number?

#random intercept
p.lme5<-lme(wing_area~cell_area,random=~1|stock,data=dat,method="ML")
summary(p.lme5)
#random intercepts and slopes
p.lme6<-lme(wing_area~cell_area,random=~1+cell_area|stock,data=dat,method="ML")
summary(p.lme6)
#compare fits
AIC(p.lme5,p.lme6);anova(p.lme5,p.lme6)
#random intercepts are better

p.lme7<-lme(wing_area~cell_number,random=~1|stock,data=dat,method="ML")
summary(p.lme7)
p.lme8<-lme(wing_area~cell_number,random=~1+cell_number|stock,data=dat,method="ML")
summary(p.lme8)
#compare fits
AIC(p.lme7,p.lme8);anova(p.lme7,p.lme8)
#random intercepts and slopes are better
# ------------------------------------------------------------------------------
# Table 2
# ------------------------------------------------------------------------------
# Females fresh mass
f.lme1<-lmer(fresh_mass_log~cell_area_log+ (1|stock),REML = TRUE, data=female)

f.lme3<-lmer(fresh_mass_log~cell_number_log+ (1|stock),REML = TRUE,data=female)

fit.list.table.2.f <- list(f.lme1,f.lme3)
fit.names.table.2.f <-c("fresh mass ~ cell area",
                        "fresh mass ~ cell number")

#compare by using AICc
fit.table.2.f<-aictab(fit.list.table.2.f,fit.names.table.2.f, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.table.2.f

# Females wing area
f.lme5<-lmer(wing_area_log~cell_area_log+ (1|stock),REML = TRUE,data=female)

f.lme7<-lmer(wing_area_log~cell_number_log+ (1|stock),REML = TRUE,data=female)

fit.list.table.2.w <- list(f.lme5,f.lme7)
fit.names.table.2.w <-c("wing area ~ cell area",
                        "wing area ~ cell number")

#compare by using AICc
fit.table.2.w<-aictab(fit.list.table.2.w,fit.names.table.2.w, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.table.2.w

# Males fresh mass
m.lme1<-lmer(fresh_mass_log~cell_area_log + (1|stock),REML = TRUE,data=male)

m.lme3<-lmer(fresh_mass_log~cell_number_log+ (1|stock),REML = TRUE,data=male)

fit.list.table.2.f <- list(m.lme1,m.lme3)
fit.names.table.2.f <-c("fresh mass ~ cell area",
                        "fresh mass ~ cell number")

#compare by using AICc
fit.table.2.f<-aictab(fit.list.table.2.f,fit.names.table.2.f, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.table.2.f


# Males wing area
m.lme5<-lmer(wing_area_log~cell_area_log+ (1|stock),REML = TRUE,data=male)

m.lme7<-lmer(wing_area_log~cell_number_log+ (1|stock),REML = TRUE,data=male)

fit.list.table.2.w <- list(m.lme5,m.lme7)
fit.names.table.2.w <-c("wing area ~ cell area",
                        "wing area ~ cell number")

#compare by using AICc
fit.table.2.w<-aictab(fit.list.table.2.w,fit.names.table.2.w, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.table.2.w


# Sex pooled fresh mass
p.lme1<-lmer(fresh_mass_log~cell_area_log+ (1|stock),REML = TRUE,data=dat)

p.lme3<-lmer(fresh_mass_log~cell_number_log+ (1|stock),REML = TRUE,data=dat)

fit.list.table.2.f <- list(p.lme1,p.lme3)
fit.names.table.2.f <-c("fresh mass ~ cell area",
                        "fresh mass ~ cell number")

#compare by using AICc
fit.table.2.f<-aictab(fit.list.table.2.f,fit.names.table.2.f, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.table.2.f


# Sex pooled wing area
p.lme5<-lmer(wing_area_log ~ cell_area_log + (1|stock),REML = TRUE,data=dat)

p.lme7<-lmer(wing_area_log ~ cell_number_log + (1|stock),REML = TRUE,data=dat)

fit.list.table.2.w <- list(p.lme5,p.lme7)
fit.names.table.2.w <-c("wing area ~ cell area",
                        "wing area ~ cell number")

#compare by using AICc
fit.table.2.w<-aictab(fit.list.table.2.w,fit.names.table.2.w, second.ord = T,sort = TRUE, digits = 3, LL=TRUE)
fit.table.2.w

# ------------------------------------------------------------------------------
# Figure 3 and Figure 4
# ------------------------------------------------------------------------------
{
# Fresh mass vs cell size in females
f.lme1<-lmer(fresh_mass_log~cell_area_log+ (1|stock),REML = TRUE, data=female)
#coef(summary(f.lme1))[ , "Estimate"]

female$pred <- predict(f.lme1,re.form=NA)
female$pred1 <- predict(f.lme1)
f.lme1 <- ggplot(female,aes(cell_area_log,fresh_mass_log))+
    #ylim(c(-0.26,0.25))+xlim(c(1.9,2.45))+
    geom_point(aes(shape= stock,color = stock))+
    scale_shape_manual(values=c(8,21:25))+
    scale_color_manual(values = c("#4DAC26","#4DAC26","#4DAC26","#4DAC26","#4DAC26","#4DAC26"))
f.lme1.plot=
    f.lme1 +   geom_line(colour="gray",aes(y=pred1,group=stock)) +
    geom_line(colour="#4DAC26",lwd=2,aes(y=pred,group=stock))+
    # xlab(expression(paste(log[10], "[cell area (",mu, m^2,")]"))) + 
    # ylab(expression(paste(log[10],"[fresh mass (mg)]"))) + 
    scale_x_continuous(name = expression(paste("cell area (",mu, m^2,")")),
                       labels = scales::math_format(10^.x),lim=c(1.9,2.45)) +
    scale_y_continuous(name = expression(paste("fresh mass (mg)")),
                       labels = scales::math_format(10^.x), lim=c(-0.26,0.25)) +
    ggtitle("Females")+
    theme(plot.title = element_text(hjust = 0.5,size = 40,face = "bold"))+
    theme_bw()+
    theme(legend.position = c(0.15,0.72), 
          legend.background = element_rect(fill = "white"),
          legend.title = element_blank())
f.lme1.plot

# Fresh mass vs cell number in females
f.lme3<-lmer(fresh_mass_log~cell_number_log+ (1|stock),REML = TRUE,data=female)

female$pred <- predict(f.lme3,re.form=NA)
female$pred1 <- predict(f.lme3)
f.lme3 <- ggplot(female,aes(cell_number_log,fresh_mass_log))+
    #ylim(c(-0.26,0.25))+xlim(c(3.7,4))+
    geom_point(aes(shape= stock,color = stock))+
    scale_shape_manual(values=c(8,21:25))+
    scale_color_manual(values = c("#4DAC26","#4DAC26","#4DAC26","#4DAC26","#4DAC26","#4DAC26"))
f.lme3.plot=
    f.lme3 +   geom_line(colour="gray",aes(y=pred1,group=stock)) +
    geom_line(colour="#4DAC26",lwd=2,aes(y=pred,group=stock))+
    # xlab(expression(paste(log[10],"[cell number]"))) + 
    # ylab(expression(paste(log[10],"[fresh mass (mg)]"))) + 
    scale_x_continuous(name = "cell number",
                       labels = scales::math_format(10^.x),lim=c(3.7,4)) +
    scale_y_continuous(name = expression(paste("fresh mass (mg)")),
                       labels = scales::math_format(10^.x), lim=c(-0.26,0.25)) +
    ggtitle("Females")+
    theme(plot.title = element_text(hjust = 0.5,size = 20,face = "bold"))+
    theme_bw() +
    theme(legend.position = c(0.85, 0.3), 
          legend.background = element_rect(fill = "white"),
          legend.title = element_blank())
f.lme3.plot

# Wing area vs cell size in females
f.lme5<-lmer(wing_area_log~cell_area_log+ (1|stock),REML = TRUE,data=female)

female$pred <- predict(f.lme5,re.form=NA)
female$pred1 <- predict(f.lme5)
f.lme5 <- ggplot(female,aes(cell_area_log,wing_area_log))+
    #ylim(c(5.8,6.3))+xlim(c(1.9,2.45))+
    geom_point(aes(shape= stock,color = stock))+
    scale_shape_manual(values=c(8,21:25))+
    scale_color_manual(values = c("#4DAC26","#4DAC26","#4DAC26","#4DAC26","#4DAC26","#4DAC26"))

f.lme5.plot=
    f.lme5 +   geom_line(colour="gray",aes(y=pred1,group=stock)) +
    geom_line(colour="#4DAC26",lwd=2,aes(y=pred,group=stock))+
    # xlab(expression(paste(log[10], "[cell area (",mu, m^2,")]"))) + 
    # ylab(expression(paste(log[10], "[wing area (",mu, m^2,")]"))) + 
    scale_x_continuous(name = expression(paste("cell area (",mu, m^2,")")),
    labels = scales::math_format(10^.x),lim=c(1.9,2.45)) +
    scale_y_continuous(name = expression(paste("fresh mass (mg)")),
                       labels = scales::math_format(10^.x), lim=c(5.8,6.3)) +
    theme(plot.title = element_text(hjust = 0.5,size = 40,face = "bold"))+
    theme_bw() +
    theme(legend.position = c(0.15,0.72), 
          legend.background = element_rect(fill = "white"),
          legend.title = element_blank())
f.lme5.plot

# Wing area vs cell number in females
f.lme7<-lmer(wing_area_log~cell_number_log+ (1|stock),REML = TRUE,data=female)

female$pred <- predict(f.lme7,re.form=NA)
female$pred1 <- predict(f.lme7)
f.lme7 <- ggplot(female,aes(cell_number_log,wing_area_log))+
    #ylim(c(5.8,6.3))+xlim(c(3.7,4))+
    geom_point(aes(shape= stock,color = stock))+
    scale_shape_manual(values=c(8,21:25))+
    scale_color_manual(values = c("#4DAC26","#4DAC26","#4DAC26","#4DAC26","#4DAC26","#4DAC26"))

f.lme7.plot=
    f.lme7 +   geom_line(colour="gray",aes(y=pred1,group=stock)) +
    geom_line(colour="#4DAC26",lwd=2,aes(y=pred,group=stock))+
    # xlab(expression(paste(log[10],"[cell number]"))) + 
    # ylab(expression(paste(log[10], "[wing area (",mu, m^2,")]"))) +
    scale_x_continuous(name = "cell number",
                       labels = scales::math_format(10^.x),lim=c(3.7,4)) +
    scale_y_continuous(name = expression(paste("wing area (",mu, m^2,")")),
                       labels = scales::math_format(10^.x), lim=c(5.8,6.3)) +
    theme(plot.title = element_text(hjust = 0.5,size = 20,face = "bold"))+
    theme_bw() +
    theme(legend.position = c(0.85, 0.3), 
          legend.background = element_rect(fill = "white"),
          legend.title = element_blank())
f.lme7.plot
# ------------------------------------------------------------------------------
# Fresh mass vs cell size in males
m.lme1<-lmer(fresh_mass_log~cell_area_log + (1|stock),REML = TRUE,data=male)

male$pred <- predict(m.lme1,re.form=NA)
male$pred1 <- predict(m.lme1)
m.lme1 <- ggplot(male,aes(cell_area_log,fresh_mass_log))+
    #ylim(c(-0.26,0.25))+xlim(c(1.9,2.45))+
    geom_point(aes(shape= stock,color = stock))+
    scale_shape_manual(values=c(8,21:25))+
    scale_color_manual(values = c("#A6761D","#A6761D","#A6761D","#A6761D","#A6761D","#A6761D"))

m.lme1.plot=
    m.lme1 +   geom_line(colour="gray",aes(y=pred1,group=stock)) +
    geom_line(colour="#A6761D",lwd=2,aes(y=pred,group=stock))+
    # xlab(expression(paste(log[10], "[cell area (",mu, m^2,")]"))) + 
    # ylab(expression(paste(log[10],"[fresh mass (mg)]"))) + 
    scale_x_continuous(name = expression(paste("cell area (",mu, m^2,")")),
                       labels = scales::math_format(10^.x),lim=c(1.9,2.45)) +
    scale_y_continuous(name = expression(paste("fresh mass (mg)")),
                       labels = scales::math_format(10^.x), lim=c(-0.26,0.25)) +
    ggtitle("Males")+
    theme(plot.title = element_text(hjust = 0.5,size = 20,face = "bold"))+
    theme_bw() +
    theme(legend.position = c(0.15,0.72), 
          legend.background = element_rect(fill = "white"),
          legend.title = element_blank())
m.lme1.plot

# Fresh mass vs cell number in males
m.lme3<-lmer(fresh_mass_log~cell_number_log+ (1|stock),REML = TRUE,data=male)

male$pred <- predict(m.lme3,re.form=NA)
male$pred1 <- predict(m.lme3)
m.lme3 <- ggplot(male,aes(cell_number_log,fresh_mass_log))+
    #ylim(c(-0.26,0.25))+xlim(c(3.7,4))+
    geom_point(aes(shape= stock,color = stock))+
    scale_shape_manual(values=c(8,21:25))+
    scale_color_manual(values = c("#A6761D","#A6761D","#A6761D","#A6761D","#A6761D","#A6761D"))

m.lme3.plot=
    m.lme3 +   geom_line(colour="gray",aes(y=pred1,group=stock)) +
    geom_line(colour="#A6761D",lwd=2,aes(y=pred,group=stock))+
    # xlab(expression(paste(log[10],"[cell number]"))) + 
    # ylab(expression(paste(log[10],"[fresh mass (mg)]"))) + 
    scale_x_continuous(name = "cell number",
                       labels = scales::math_format(10^.x),lim=c(3.7,4)) +
    scale_y_continuous(name = expression(paste("fresh mass (mg)")),
                       labels = scales::math_format(10^.x), lim=c(-0.26,0.25)) +
    ggtitle("Males")+
    theme(plot.title = element_text(hjust = 0.5,size = 20,face = "bold"))+
    theme_bw() +
    theme(legend.position = c(0.85, 0.3), 
          legend.background = element_rect(fill = "white"),
          legend.title = element_blank())
m.lme3.plot

# Wing area vs cell size in males
m.lme5<-lmer(wing_area_log~cell_area_log+ (1|stock),REML = TRUE,data=male)

male$pred <- predict(m.lme5,re.form=NA)
male$pred1 <- predict(m.lme5)
m.lme5 <- ggplot(male,aes(cell_area_log,wing_area_log))+
    #ylim(c(5.8,6.3))+xlim(c(1.9,2.45))+
    geom_point(aes(shape= stock,color = stock))+
    scale_shape_manual(values=c(8,21:25))+
    scale_color_manual(values = c("#A6761D","#A6761D","#A6761D","#A6761D","#A6761D","#A6761D"))

m.lme5.plot=
    m.lme5 +   geom_line(colour="gray",aes(y=pred1,group=stock)) +
    geom_line(colour="#A6761D",lwd=2,aes(y=pred,group=stock))+
    # xlab(expression(paste(log[10], "[cell area (",mu, m^2,")]"))) + 
    # ylab(expression(paste(log[10], "[wing area (",mu, m^2,")]"))) + 
    scale_x_continuous(name = expression(paste("cell area (",mu, m^2,")")),
                       labels = scales::math_format(10^.x),lim=c(1.9,2.45)) +
    scale_y_continuous(name = expression(paste("wing area (",mu, m^2,")")),
                       labels = scales::math_format(10^.x), lim=c(5.8,6.3)) +
    theme(plot.title = element_text(hjust = 0.5,size = 20,face = "bold"))+
    theme_bw() +
    theme(legend.position = c(0.15,0.72), 
          legend.background = element_rect(fill = "white"),
          legend.title = element_blank())
m.lme5.plot

# Wing area vs cell number in males
m.lme7<-lmer(wing_area_log~cell_number_log+ (1|stock),REML = TRUE,data=male)

male$pred <- predict(m.lme7,re.form=NA)
male$pred1 <- predict(m.lme7)
m.lme7 <- ggplot(male,aes(cell_number_log,wing_area_log))+
    #ylim(c(5.8,6.3))+xlim(c(3.7,4))+
    geom_point(aes(shape= stock,color = stock))+
    scale_shape_manual(values=c(8,21:25))+
    scale_color_manual(values = c("#A6761D","#A6761D","#A6761D","#A6761D","#A6761D","#A6761D"))

m.lme7.plot=
    m.lme7 +   geom_line(colour="gray",aes(y=pred1,group=stock)) +
    geom_line(colour="#A6761D",lwd=2,aes(y=pred,group=stock))+
    # xlab(expression(paste(log[10],"[cell number]"))) + 
    # ylab(expression(paste(log[10], "[wing area (",mu, m^2,")]"))) +
    scale_x_continuous(name = "cell number",
                       labels = scales::math_format(10^.x),lim=c(3.7,4)) +
    scale_y_continuous(name = expression(paste("wing area (",mu, m^2,")")),
                       labels = scales::math_format(10^.x), lim=c(5.8,6.3)) +
    theme(plot.title = element_text(hjust = 0.5,size = 20,face = "bold"))+
    theme_bw() +
    theme(legend.position = c(0.85, 0.3), 
          legend.background = element_rect(fill = "white"),
          legend.title = element_blank())
m.lme7.plot

# ------------------------------------------------------------------------------
# Fresh mass vs cell size in sex pooled
p.lme1<-lmer(fresh_mass_log~cell_area_log + (1|stock),REML = TRUE,data=dat)

dat$pred <- predict(p.lme1,re.form=NA)
dat$pred1 <- predict(p.lme1)
p.lme1 <- ggplot(dat,aes(cell_area_log,fresh_mass_log))+
    #ylim(c(-0.26,0.25))+xlim(c(1.9,2.45))+
    geom_point(aes(shape= stock,color = stock))+
    scale_shape_manual(values=c(8,21:25))+
    scale_color_manual(values = c("#555555","#555555","#555555","#555555","#555555","#555555"))

p.lme1.plot=
    p.lme1 +   geom_line(colour="gray",aes(y=pred1,group=stock)) +
    geom_line(colour="#555555",lwd=2,aes(y=pred,group=stock))+
    # xlab(expression(paste(log[10], "[cell area (",mu, m^2,")]"))) + 
    # ylab(expression(paste(log[10],"[fresh mass (mg)]"))) + 
    scale_x_continuous(name = expression(paste("cell area (",mu, m^2,")")),
                       labels = scales::math_format(10^.x),lim=c(1.9,2.45)) +
    scale_y_continuous(name = "fresh mass (mg)",
                       labels = scales::math_format(10^.x), lim=c(-0.26,0.25)) +
    ggtitle("Sex pooled")+
    theme(plot.title = element_text(hjust = 0.5,size = 40,face = "bold"))+
    theme_bw() +
    theme(legend.position = c(0.15,0.72), 
          legend.background = element_rect(fill = "white"),
          legend.title = element_blank())
p.lme1.plot

# Fresh mass vs cell number in sex pooled
p.lme3<-lmer(fresh_mass_log~cell_number_log+ (1|stock),REML = TRUE,data=dat)

dat$pred <- predict(p.lme3,re.form=NA)
dat$pred1 <- predict(p.lme3)
p.lme3 <- ggplot(dat,aes(cell_number_log,fresh_mass_log))+
    #ylim(c(-0.26,0.25))+xlim(c(3.7,4))+
    geom_point(aes(shape= stock,color = stock))+
    scale_shape_manual(values=c(8,21:25))+
    scale_color_manual(values = c("#555555","#555555","#555555","#555555","#555555","#555555"))

p.lme3.plot=
    p.lme3 +   geom_line(colour="gray",aes(y=pred1,group=stock)) +
    geom_line(colour="#555555",lwd=2,aes(y=pred,group=stock))+
    # xlab(expression(paste(log[10],"[cell number]"))) + 
    # ylab(expression(paste(log[10],"[fresh mass (mg)]"))) + 
    scale_x_continuous(name = "cell number",
                       labels = scales::math_format(10^.x),lim=c(3.7,4)) +
    scale_y_continuous(name = "fresh mass (mg)",
                       labels = scales::math_format(10^.x), lim=c(-0.26,0.25)) +
    ggtitle("Females")+
    ggtitle("Sex pooled")+
    theme(plot.title = element_text(hjust = 0.5,size = 20,face = "bold"))+
    theme_bw() +
    theme(legend.position = c(0.85, 0.3), 
          legend.background = element_rect(fill = "white"),
          legend.title = element_blank())
p.lme3.plot

# Wing area vs cell size in sex pooled
p.lme5<-lmer(wing_area_log~cell_area_log+ (1|stock),REML = TRUE,data=dat)

dat$pred <- predict(p.lme5,re.form=NA)
dat$pred1 <- predict(p.lme5)
p.lme5 <- ggplot(dat,aes(cell_area_log,wing_area_log))+
    #ylim(c(5.8,6.3))+xlim(c(1.9,2.45))+
    geom_point(aes(shape= stock,color = stock))+
    scale_shape_manual(values=c(8,21:25))+
    scale_color_manual(values = c("#555555","#555555","#555555","#555555","#555555","#555555"))

p.lme5.plot=
    p.lme5 +   geom_line(colour="gray",aes(y=pred1,group=stock)) +
    geom_line(colour="#555555",lwd=2,aes(y=pred,group=stock))+
    # xlab(expression(paste(log[10], "[cell area (",mu, m^2,")]"))) + 
    # ylab(expression(paste(log[10], "[wing area (",mu, m^2,")]"))) + 
    scale_x_continuous(name = expression(paste("cell area (",mu, m^2,")")),
                       labels = scales::math_format(10^.x),lim=c(1.9,2.45)) +
    scale_y_continuous(name = expression(paste("wing area (",mu, m^2,")")),
                       labels = scales::math_format(10^.x), lim=c(5.8,6.3)) +
    theme(plot.title = element_text(hjust = 0.5,size = 20,face = "bold"))+
    theme_bw() +
    theme(legend.position = c(0.15,0.72), 
          legend.background = element_rect(fill = "white"),
          legend.title = element_blank())
p.lme5.plot

# Wing area vs cell number in sex pooled
p.lme7<-lmer(wing_area_log~cell_number_log + (1|stock),REML = TRUE,data=dat)

dat$pred <- predict(p.lme7,re.form=NA)
dat$pred1 <- predict(p.lme7)
p.lme7 <- ggplot(dat,aes(cell_number_log,wing_area_log))+
    #ylim(c(5.8,6.3))+xlim(c(3.7,4))+
    geom_point(aes(shape= stock,color = stock))+
    scale_shape_manual(values=c(8,21:25))+
    scale_color_manual(values = c("#555555","#555555","#555555","#555555","#555555","#555555"))

p.lme7.plot=
    p.lme7 +   geom_line(colour="gray",aes(y=pred1,group=stock)) +
    geom_line(colour="#555555",lwd=2,aes(y=pred,group=stock))+
    # xlab(expression(paste(log[10],"[cell number]"))) + 
    # ylab(expression(paste(log[10], "[wing area (",mu, m^2,")]"))) +
    scale_x_continuous(name = "cell number",
                       labels = scales::math_format(10^.x),lim=c(3.7,4)) +
    scale_y_continuous(name = expression(paste("wing area (",mu, m^2,")")),
                       labels = scales::math_format(10^.x), lim=c(5.8,6.3)) +
    theme(plot.title = element_text(hjust = 0.5,size = 20,face = "bold"))+
    theme_bw() +
    theme(legend.position = c(0.85, 0.3), 
          legend.background = element_rect(fill = "white"),
          legend.title = element_blank())
p.lme7.plot
# ------------------------------------------------------------------------------
setwd("../Outputs/")#Set directory
#Figure 3
Figure_3<-plot_grid(f.lme1.plot,m.lme1.plot,p.lme1.plot,
                    f.lme5.plot,m.lme5.plot,p.lme5.plot,
                    labels=c("A","B","C",
                              "D","E","F")
                    ,nrow = 2,ncol = 3,label_size=15)

#Figure 4
Figure_4<-plot_grid(f.lme3.plot,m.lme3.plot,p.lme3.plot,
                    f.lme7.plot,m.lme7.plot,p.lme7.plot,
                    labels=c("A","B","C",
                             "D","E","F")
                    ,nrow = 2,ncol = 3,label_size=15)

##Store Plots
ggsave('3.2.1. Figure 3.pdf',Figure_3,width=12,height=8)
ggsave('3.2.1. Figure 3.png',Figure_3,width=12,height=8)
ggsave('3.2.2. Figure 4.pdf',Figure_4,width=12,height=8)
ggsave('3.2.2. Figure 4.png',Figure_4,width=12,height=8)
}
# ------------------------------------------------------------------------------
# saving session information with all packages versions for reproducibility purposes
sink("../Outputs/3.1.1. Data analyses_Table_1_and_Figure_3_R_session.txt")
sessionInfo()
sink()
# ------------------------------------------------------------------------------
################################################################################
############################ END OF SCRIPT #####################################
################################################################################
