==> R/4.0. Data analyses_Table_3_and_Figure_5.R <==
# ------------------------------------------------------------------------------
# Data generated by Félix P Leiva (e-mail: felixpleiva@gmail.com)
# Script created by Félix P Leiva (e-mail: felixpleiva@gmail.com)
# ------------------------------------------------------------------------------
# When using the data or code from this manuscript, please cite it as:

#Leiva FP, Boerrigter JGJ & Verberk WCEP. (2023). The role of cell size in
#shaping responses to oxygen and temperature in fruit flies. Zenodo.
#https://doi.org/10.5281/zenodo.5949048.
# ------------------------------------------------------------------------------
# Cleaning working space
rm(list=ls())
today<-format(Sys.Date(),"%Y%m%d")
# ------------------------------------------------------------------------------
# set the working directory to the folder containing this script:
setwd(dirname(rstudioapi::getActiveDocumentContext()$path))

# check directory
getwd()
# ------------------------------------------------------------------------------
#Libraries
library(lme4)
library(car)
library(visreg)
library(AICcmodavg)
# ------------------------------------------------------------------------------
# create a folder for the outputs produced by running this script (if it doesn't
# already exist)
if (!file.exists("../Outputs")) dir.create("../Outputs")
# '../' goes up one level from the current working directory, so this creates
# the 'Outputs' folder just outside the 'R' folder
# ------------------------------------------------------------------------------
# load data
dat<-read.csv("../Outputs/1.1.1. Data to fit the models.csv")
# ------------------------------------------------------------------------------
# Because cell size was the main cause of changes in body mass and wing area, we
# test if such scaling relationships between depend on oxygen or temperature
str(dat)
dat$cell_area_cat<-as.factor(dat$cell_area_cat)
dat$stock<-as.factor(dat$stock)
dat$temperature<-as.numeric(dat$temperature)
dat$oxygen<-as.numeric(dat$oxygen)
dat$sex<-as.factor(dat$sex)
dat$ID_fly<-as.factor(dat$ID_fly)

#model
fit1<-lmer(fresh_mass_log ~ temperature * cell_area_log + oxygen * cell_area_log + sex + (1|stock),REML = TRUE, data=dat)
summary(fit1)

# Fixed effects:
#                        Estimate Std. Error t value
# (Intercept)               -0.922156   0.342771  -2.690
# temperature               -0.032520   0.012895  -2.522
# cell_area_log              0.424838   0.153571   2.766
# oxygen                     0.056222   0.007939   7.082
# sexmale                   -0.145913   0.004645 -31.413
# temperature:cell_area_log  0.014117   0.005848   2.414
# cell_area_log:oxygen      -0.024029   0.003598  -6.678
#
Anova(fit1)
# Response: fresh_mass_log
#                               Chisq Df Pr(>Chisq)    
# temperature                 5.4334  1    0.01976 *  
# cell_area_log              56.8101  1  4.800e-14 ***
# oxygen                    110.5958  1  < 2.2e-16 ***
# sex                       986.7997  1  < 2.2e-16 ***
# temperature:cell_area_log   5.8273  1    0.01578 *  
# cell_area_log:oxygen       44.5903  1  2.429e-11 ***
# ---
# Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
################################################################################
# Figure 5
################################################################################
setwd("../Outputs/")#Set directory
{
#pdf("4.2.1. Figure 5.pdf",width = 10,height = 5,useDingbats = FALSE)
png("4.2.1. Figure 5.png",width = 10,height = 5,units = "in",res = 600)
par(mfrow=c(1,2),tcl=-0.4, family="serif",omi=c(0,0,0,0.1))

# labels for y-axis
labs_mass = c("0.4","0.6","0.8","1.0","1.2")
xs_mass = log10(c(0.4,0.6,0.8,1.0,1.2))
# labels for x-axis
labs_cell = c("80","105","135","170","220")
xs_cell = log10(c(80,105,135,170,220))

# under hypoxia
par(mai=c(0.85,0.87,0.1,0.1))
visreg(fit1,"cell_area_log",by="temperature",cond=list(oxygen=10),overlay=TRUE,
       print.cond=TRUE,legend = FALSE,
       points=list(col=c("#2c7fb870", "#de2d2670"),pch=16,cex=1.4),
       line=list(col=c("#2c7fb8", "#de2d26"), lwd=4),
       #xlab=expression(paste(log[10], "[cell area (",mu, m^2,")]")),
       xlab=expression(paste("cell area (",mu, m^2,")")),
       #ylab=expression(paste(log[10],"[fresh mass (mg)]")),
       ylab="fresh mass (mg)",
       yaxt="n",
       xaxt="n",
       ylim=c(-0.4,0.1),
       xlim=c(1.9,2.4),
       cex.axis=1.2,cex.lab=1.4)

#update axis
axis(2, at=xs_mass, labels=labs_mass, cex.axis=1.2,las=1)
axis(1, at=xs_cell, labels=labs_cell, cex.axis=1.2,las=1)

## Draw a polygon defining an area on the graph
xx = c(1.8815,1.8815,2.4185,2.4185)
yy = c(0.05,0.119,0.119,0.05)
polygon(xx,yy, col = 'grey90', border = NA)

legend("bottomright", legend=c("17°C", "25°C"),
       lty = 1,lwd=4, bty="n",
       col=c("#2c7fb8", "#de2d26"),cex=1.4)
text(2.15,0.0845,"Hypoxia (10 kPa)",font=2,cex=1.5)

# under normoxia
par(mai=c(0.85,0.87,0.1,0.1))
visreg(fit1,"cell_area_log",by="temperature",cond=list(oxygen=21),overlay=TRUE,
       print.cond=TRUE,legend = FALSE,
       points=list(col=c("#2c7fb870", "#de2d2670"),pch=16,cex=1.4),
       line=list(col=c("#2c7fb8", "#de2d26"), lwd=4),
       #xlab=expression(paste(log[10], "[cell area (",mu, m^2,")]")),
       xlab=expression(paste("cell area (",mu, m^2,")")),
       #ylab=expression(paste(log[10],"[fresh mass (mg)]")),
       ylab="fresh mass (mg)",
       yaxt="n",
       xaxt="n",
       ylim=c(-0.4,0.1),
       xlim=c(1.9,2.4),
       cex.axis=1.2,cex.lab=1.4)

#update axis
axis(2, at=xs_mass, labels=labs_mass, cex.axis=1.2,las=1)
axis(1, at=xs_cell, labels=labs_cell, cex.axis=1.2,las=1)

## Draw a polygon defining an area on the graph
xx = c(1.8815,1.8815,2.4185,2.4185)
yy = c(0.05,0.119,0.119,0.05)
polygon(xx,yy, col = 'grey90', border = NA)
legend("bottomright", legend=c("17°C", "25°C"),
       lty = 1,lwd=4, bty="n",
       col=c("#2c7fb8", "#de2d26"),cex=1.4)
text(2.15,0.0845,"Normoxia (21 kPa)",font=2,cex=1.5)

dev.off()
}
# ------------------------------------------------------------------------------
# saving session information with all packages versions for reproducibility purposes
sink("../Outputs/4.1.1. Data analyses_Tables_3_4_and_Figure_5_R_session.txt")
sessionInfo()
sink()
# ------------------------------------------------------------------------------
################################################################################
############################ END OF SCRIPT #####################################
################################################################################
