==> R/5.0. Data analyses_Figure_6.R <==
# ------------------------------------------------------------------------------
# Data generated by Félix P Leiva (e-mail: felixpleiva@gmail.com)
# Script created by Félix P Leiva (e-mail: felixpleiva@gmail.com)
# ------------------------------------------------------------------------------
# When using the data or code from this manuscript, please cite it as:

#Leiva FP, Boerrigter JGJ & Verberk WCEP. (2023). The role of cell size in
#shaping responses to oxygen and temperature in fruit flies. Zenodo.
#https://doi.org/10.5281/zenodo.5949048.
# ------------------------------------------------------------------------------
# Cleaning working space
rm(list=ls())
today<-format(Sys.Date(),"%Y%m%d")
# ------------------------------------------------------------------------------
# set the working directory to the folder containing this script:
setwd(dirname(rstudioapi::getActiveDocumentContext()$path))

# check directory
getwd()
# ------------------------------------------------------------------------------
#Libraries
library(lme4)
library(car)
library(visreg)
library(AICcmodavg)
# ------------------------------------------------------------------------------
# create a folder for the outputs produced by running this script (if it doesn't
# already exist)
if (!file.exists("../Outputs")) dir.create("../Outputs")
# '../' goes up one level from the current working directory, so this creates
# the 'Outputs' folder just outside the 'R' folder
# ------------------------------------------------------------------------------
# load data
dat<-read.csv("../Outputs/1.1.1. Data to fit the models.csv")

# ------------------------------------------------------------------------------
# filter by sex
female=subset(dat,sex!="male")
male=subset(dat,sex=="male")
female2<-subset(female,female$oxygen==21&female$temperature==25)
male2<-subset(male,male$oxygen==21&male$temperature==25)
dat2<-subset(male,dat$oxygen==21&dat$temperature==25)

# ------------------------------------------------------------------------------
################################################################################
# Fitting the model with relative values of fresh mass and cell area
fit1<-lmer(fresh_mass_delta ~ temperature * cell_area_delta + oxygen * cell_area_delta + sex + (1|stock),REML = FALSE, data=dat)
summary(fit1); Anova(fit1)

# Table S4
Table_S4 <- capture.output(summary(fit1))
setwd("../Outputs/")
write.table(Table_S4,"5.3.1. Table_S4_Estimates of the model used to plot Figure 6.txt",row.names = F,quote = F)

################################################################################
# Figure 6
################################################################################
{
pdf("5.2.1. Figure 6.pdf",width = 10,height = 5,useDingbats = FALSE)
#png("5.2.1. Figure 6.png",width = 10,height = 5,units = "in",res = 600)
par(mfrow=c(1,2),tcl=-0.4, family="serif",omi=c(0,0,0,0))

# under hypoxia
par(mai=c(0.82,0.82,0.2,0.2))
visreg(fit1,"cell_area_delta",by="temperature",cond=list(oxygen=10),overlay=TRUE,
   print.cond=TRUE,legend = FALSE,
       points=list(col=c("#2c7fb870", "#de2d2670"),pch=16,cex=1.4),
       line=list(col=c("#2c7fb8", "#de2d26"), lwd=4),
       xlab="standardised cell area (%)", ylab="standardised fresh mass (%)",
       ylim=c(-50,50),xlim=c(-50,50),
       xaxs = "i",yaxs = "i",#force origin of our x- and y-axes
       cex.axis=1.5,cex.lab=1.5)
abline(h=0,lty=2,lwd=2)
abline(v=0,lty=2,lwd=2)
## Draw a polygon defining an area on the graph
xx = c(-49.8,-49.8,49.8,49.8)
yy = c(35,49.8,49.8,35)
polygon(xx,yy, col = 'grey90', border = NA)
legend("bottomright", legend=c("17°C", "25°C"),
       lty = 1,lwd=4, bty="n",
       col=c("#2c7fb8", "#de2d26"),cex=1.4)
text(0,42.5,"Hypoxia (10 kPa)",font=2,cex=1.5)

# under normoxia
par(mai=c(0.82,0.82,0.2,0.2))
visreg(fit1,"cell_area_delta",by="temperature",cond=list(oxygen=21),overlay=TRUE,
       print.cond=TRUE,legend = FALSE,
       points=list(col=c("#2c7fb870", "#de2d2670"),pch=16,cex=1.4),
       line=list(col=c("#2c7fb8", "#de2d26"), lwd=4),
       xlab="standardised cell area (%)", ylab="standardised fresh mass (%)",
       ylim=c(-50,50),xlim=c(-50,50),
       xaxs = "i",yaxs = "i",#force origin of our x- and y-axes
       cex.axis=1.5,cex.lab=1.5)
abline(h=0,lty=2,lwd=2)
abline(v=0,lty=2,lwd=2)
## Draw a polygon defining an area on the graph
xx = c(-49.8,-49.8,49.8,49.8)
yy = c(35,49.8,49.8,35)
polygon(xx,yy, col = 'grey90', border = NA)
legend("bottomright", legend=c("17°C", "25°C"),
       lty = 1,lwd=4, bty="n",
       col=c("#2c7fb8", "#de2d26"),cex=1.4)
text(0,42.5,"Normoxia (21 kPa)",font=2,cex=1.5)

dev.off()
}
# ------------------------------------------------------------------------------
# saving session information with all packages versions for reproducibility purposes
sink("../Outputs/5.1.1. Data analyses_Figure_6_R_session.txt")
sessionInfo()
sink()
# ------------------------------------------------------------------------------
################################################################################
############################ END OF SCRIPT #####################################
################################################################################
